[TRACE] Fixed a warning

+++ src/emu/x64run_private.c
@@ -1124,7 +1124,7 @@ void PrintTrace(x64emu_t* emu, uintptr_t ip, int dynarec)
                 printFunctionAddr(nextaddr, "=> ");
             } else if(peek==0xFF) {
                 if(PK(1)==0x25) {
-                    uintptr_t nextaddr = is32bits?(*(uint32_t*)PK32(2)):(*(uintptr_t*)(ip + 6 + PK32(2)));
+                    uintptr_t nextaddr = is32bits?(*(uint32_t*)(uintptr_t)PK32(2)):(*(uintptr_t*)(ip + 6 + PK32(2)));
                     if(!printFunctionAddr(nextaddr, "=> "))
                         printf_log(LOG_NONE, " => %p", (void*)nextaddr);
                 } else if((PK(1)==0x14) && (PK(2)==0x25)) {
